✨ feat(auth, rmq): Implement user authentication with email confirmation via RabbitMQ. ⬆️ Updates Go dependencies; ♻️ refactors generics and user roles; 🧱 establishes RabbitMQ infrastructure.

+++ internal/config/schema.go
@@ -32,7 +32,7 @@ type HTTPConfig struct {
 	Timeout   time.Duration `koanf:"timeout"`
 	Prefork   bool          `koanf:"prefork"`
 	CORS      CORSConfig    `koanf:"cors"`
-	BodyLimit int           `koanf:"bodylimit"`
+	BodyLimit int64         `koanf:"bodylimit"`
 	TLS       TLSConfig     `koanf:"tls"`
 }
